+++ graph_generation/path_downloader.py
@@ -7,7 +7,7 @@ from pyproj import Transformer
 
 
 class PathDataProcessor:
-    def __init__(self, input_data_path="data/cumbria_full.osm.pbf", output_pickle_path="Pathfinding/new_path_graph.pkl", target_epsg=27700):
+    def __init__(self, input_data_path="data/cumbria_full.osm.pbf", output_pickle_path="graph_generation/unpopulated_graph.pkl", target_epsg=27700):
         self.input_data_path = input_data_path
         self.output_pickle_path = output_pickle_path
         self.target_epsg = target_epsg
