+++ src/app.py
@@ -1,43 +1,37 @@
 #region IMPORTS
 
 # Standard Library Imports 
-import io
 import json
 import math
 import os
-import sys
 import traceback
 from datetime import datetime
-from typing import Any, Optional
+
 
 # Third-Party Libraries
-import gpxpy
-import gpxpy.gpx
 import requests
-from fastkml.kml import KML
-from fitparse import FitFile
 from flask import (
     Flask,
     jsonify,
     redirect,
     render_template,
     request,
-    send_file,
     session,
     url_for,
 )
-from flask_limiter import Limiter
-from flask_limiter.util import get_remote_address
-from sqlalchemy.exc import IntegrityError
-from sqlmodel import Session, delete, select
-from werkzeug.security import check_password_hash, generate_password_hash
+from sqlmodel import Session, select
 
 # Local Modules
 from config import Config
 from db import engine
-from models import ActionLog, BetaCode, Point, Route, Settings, User
-from Nodefinder import NodeFinder
+from models import Point, Route
 from constants import DEFAULT_CENTRE
+from extensions import limiter, service, log_action, get_current_user
+
+from Routes.routes_api import route_api_bp
+from Points.points_api import point_api_bp
+from Auth.auth_api import auth_api_bp
+from Settings.settings_api import settings_api_bp
 
 #endregion
 
@@ -48,6 +42,15 @@ app = Flask(
     template_folder='../templates', 
     static_folder='../static'
 )
+
+# Binding limiter to specific app instance
+limiter.init_app(app)
+
+# Registering Blueprints
+app.register_blueprint(route_api_bp)
+app.register_blueprint(point_api_bp)
+app.register_blueprint(auth_api_bp)
+app.register_blueprint(settings_api_bp)
 
 # Configuration Keys
 app.secret_key = Config.SECRET_KEY
@@ -60,120 +63,21 @@ if os.environ.get("FLASK_ENV") == "development":
 
 app.config["TEMPLATES_AUTO_RELOAD"] = True
 
-# Session and Security Settings 
+# Session and Security Settings
 
+# This keeps Secure cookies on in production, but disable them in local HTTP development so logins persist
+session_cookie_secure = os.environ.get("FLASK_ENV") != "development"
 
 app.config.update(
-    # TODO: CHANGE TO TRUE BEFORE PRODUCTION (Enforces HTTPS)
-    SESSION_COOKIE_SECURE=True,      
+    SESSION_COOKIE_SECURE=session_cookie_secure,
     SESSION_COOKIE_HTTPONLY=True,    # Prevents JavaScript access (XSS mitigation)
     SESSION_COOKIE_SAMESITE='Lax',   # CSRF protection
-    PERMANENT_SESSION_LIFETIME=3600  # Expires sessions after 1 hour (overrides the 30m above)
-)
-
-# Rate Limiter Setup
-limiter = Limiter(
-    get_remote_address,
-    app=app,
-    default_limits=["200 per day", "500 per hour"],
-    storage_uri="memory://",
+    PERMANENT_SESSION_LIFETIME=3600  # Expires sessions after 1 hour
 )
 
 #endregion
 
-# ==========================================
-# 3. Global Constants / Application Data
-# ==========================================
-
-#region ACTION LOGGING
-
-def log_action(
-    action: str, 
-    outcome: bool, 
-    info: Optional[Any] = None, 
-    duration_ms: Optional[int] = None,
-    code: Optional[str] = None
-) -> None:
-    """
-    Records an application event or metric into the central ActionLog table.
-
-    This function operates entirely within its own isolated database session 
-    to guarantee that transaction failures in the main application flow do 
-    not disrupt or prevent the creation of the event log. 
-
-    Args:
-        - action (str): The name of the event or endpoint (e.g., 'pathfind_request').
-
-        - outcome (bool): True if the operation succeeded, False if it failed.
-
-        - info (Any, optional): Contextual metadata. Accepts raw exceptions (which 
-            are stringified) or dictionaries/lists (which are automatically 
-            serialized into JSON strings). Defaults to None.
-
-        - duration_ms (int, optional): Performance metric representing execution 
-            time strictly formatted as milliseconds. Defaults to None.
-
-        - code (str, optional): An application-specific identifier 
-            (e.g., 'FAILED_SAVE_ROUTE').
-            Defaults to None.
-
-    Returns:
-        None
-
-    Raises:
-        Does not propagate exceptions. Internal failures (e.g., complete database 
-        unreachability) are caught silently and dumped safely to system standard error logs.
-    """
-    with Session(engine) as log_db:
-        try:
-            # This handles info parsing safely depending on data type, mainly a guard clause as most info (if not all) passed in is a string
-            processed_info = None
-            if info is not None:
-                if isinstance(info, Exception):
-                    processed_info = str(info)[:2000]
-                elif isinstance(info, (dict, list)):
-                    processed_info = json.dumps(info)
-                else:
-                    processed_info = str(info)[:2000]
-
-            # This builds the record / row
-            log_entry = ActionLog(
-                action=action,
-                outcome=outcome,
-                info=processed_info,
-                duration_ms=duration_ms,
-                error_code=code
-            )
-            
-            log_db.add(log_entry)
-            log_db.commit()
-            
-        except Exception as log_err:
-            # Fallback if the database becomes entirely unreachable
-            print(f"CRITICAL: Failed to write to ActionLog table: {log_err}", file=sys.stderr)
-            print(traceback.format_exc())
-
-#endregion
-
-
-# useful helper for getting the current user during route and point creation as well as logging in and out 
-def get_current_user():
-    username = session.get("username")
-    print(username)
-    if not username:
-        print("no username detected")
-        return None
-    with Session(engine) as db:
-        return db.exec(select(User).where(User.username == username)).first()
-
-def is_beta_code_validated():
-    beta_code = session.get("beta_code")
-    print(beta_code)
-    if not beta_code:
-        print("No beta code detected")
-        return False
-    else:
-        return True
+#region TEMPLATE FILTERS
 
 @app.template_filter('format_elevation')
 def format_elevation(elevation_gain):
@@ -202,6 +106,8 @@ def strftime_filter(date, format: str):
         date = datetime.isoformat(date)
     return date.strftime(format)
 
+#endregion
+
 
 #region ERROR HANDLER RENDER_TEMPLATES()
 
@@ -219,1275 +125,10 @@ def page_error_500(error):
 
 #endregion
 
-# region AUTH FLASK ROUTES
-
-# BETA CODE VALIDATION
-@app.route("/validate-beta-code", methods=['POST'])
-def validate_beta_code():
-    data = request.get_json()
-    code = data.get("beta_code", "").strip()
-
-    if not code:
-        return jsonify({"success": False, "message": "Beta code not received"})
-    
-    with Session(engine) as db:
-        db_code = db.exec(
-            select(BetaCode)
-            .where(BetaCode.code == code)
-        ).first()
-
-        if not db_code:
-            return jsonify({"success": False, "message": "Could not find beta code"})
-    
-    session["beta_code"] = code
-    session.permanent = True
-
-    return jsonify({"success": True, "message": "Beta Code validation was successful"})
-        
-
-#region LOGGING IN AND OUT
-
-@app.route("/login", methods=["POST"])
-@limiter.limit("10 per minute")
-def login():
-    try:
-
-        data = request.get_json()
-        username = data.get("username", "").strip()
-        password = data.get("password", "")
-
-        if not username or not password:
-            return jsonify({"success": False, "message": "Username and Password are required"})
-
-        with Session(engine) as db:
-            user = db.exec(
-                select(User).where(User.username == username)
-            ).first()
-
-        if user and check_password_hash(user.password_hashed, password):
-            session["username"] = username
-            session["preferred_name"] = user.preferred_name if user.preferred_name else user.username
-            print(session["username"])
-            print(session["preferred_name"])
-            session.permanent = True  # Make session respect PERMANENT_SESSION_LIFETIME
-            return jsonify({"success": True, "message": "Successfully logged in"})
-        if user is None or not check_password_hash(user.password_hashed, password):
-            return jsonify({"success": False, "message": "Username and/or Password are incorrect"})
-    except Exception as e:
-
-        short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-
-        log_action('Login', False, short_traceback, None, 'LOGIN')
-
-        return jsonify({"success": False, "message": "Sorry, there was an unexpected error whilst logging in."}), 500 
-
-@app.route("/logout", methods=["POST"])
-@limiter.limit("10 per minute")
-def logout():
-    username = session.get("username", "user")
-    session.pop('username', None)
-    return jsonify({"success": True, "message": f"Sucessfully logged out of {username}"})
-
-#endregion
-    
-# region REGISTERING 
-@app.route("/registering", methods=["POST"])
-@limiter.limit("10 per minute")
-def registering():
-    with Session(engine) as db:
-        try:
-            data = request.get_json()
-            username = data.get("username", "").strip()
-            p1 = data.get("password1", "")
-            p2 = data.get("password2", "")
-            preferred_name = data.get("preferred_name").strip()
-
-            if not username or len(username) <= 7:
-                return jsonify({"success": False, "message": "Username must have at least 8 characters"})
-            
-            if p1 != p2:
-                return jsonify({"success": False, "message": "The passwords must match each other"})
-            
-            if len(p1) <= 11:
-                return jsonify({"success": False, "message": "Passwords must have at least 12 characters"})
-            
-            has_digit = any(char.isdigit() for char in p1)
-            if not has_digit:
-                return jsonify({"success" : False, "message" : "Passwords must have at least one numerical digit"})
-
-            special_characters = ["@", "#", "$", "%", "^", "&", "*", "(", ")", "_", "+", "-", "=", "[", "]", "{", "}", "|", ";", ":", ",", ".", "<", ">", "?", "/"]
-            
-            if not any(character in p1 for character in special_characters):
-                return jsonify({"success": False, "message": "Passwords must have at least one special character"})
-
-            
-            existing_user = db.exec(
-                select(User).where(User.username == username)
-            ).first()
-
-        
-            if existing_user:
-                return jsonify({"success": False, "message": "Someone has already chosen this username"})
-            
-            new_user = User(
-                username = username,
-                preferred_name = preferred_name,
-                password_hashed = generate_password_hash(p1)
-            )
-
-            db.add(new_user)
-            db.commit()
-            return jsonify({"success": True, "message": "Successfully registered"})
-                
-        except Exception as error:
-            db.rollback()
-
-            short_traceback = "".join(traceback.format_exception_only(type(error), error)).strip()
-
-            log_action('Registering', False, short_traceback, None, 'FAILED_REGISTRATION')
-
-            return jsonify({"success": False, "message": "There was an unexpected error with our database"})
-
-#endregion
-
-
-
-# region DELETING ACCOUNT
-@app.route("/delete_account", methods=["POST"])
-@limiter.limit("10 per minute")
-def delete_account():
-    with Session(engine) as db:
-
-       
-        user = get_current_user()
-
-        if user:
-            try:
-                
-                db.delete(user)
-
-                db.exec(
-                    delete(Route)
-                    .where(Route.user_id == user.id)
-                )
-                db.exec(
-                    delete(Point)
-                    .where(Point.user_id == user.id)
-                )
-                
-                db.commit()
-                session.pop('username', None)
-                return jsonify({"success": True, "message": "Successfully deleted your account"})
-            except Exception as e:
-                db.rollback()
-
-                short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-
-                log_action('Deleting Account', False, short_traceback, None, 'FAILED_ACCOUNT_DELETION')
-
-                return jsonify({"success": False, "message": "Could not delete your account, try again later. "})
-    return jsonify({"success": False, "message": "Could not delete your account, try again later. "})
-
-#endregion
-
-#endregion
-
-#region SETTING ROUTES
-
-@app.route("/get_settings", methods=["GET"])
-def get_settings():
-    user = get_current_user()
-    if not user:
-        return jsonify({"success": False, "message": "Error: no user found"}), 401
-
-    try:
-        with Session(engine) as db:
-
-            # all setting records are queried
-            existing_records = db.exec(
-                select(Settings)
-                .where(Settings.user_id == user.id)
-            ).all()
-
-            # makes key value pairs in the structure (setting: user_choice)
-            settings_payload = {record.key: record.value for record in existing_records}
-            
-            # returns data in a valid format
-            return jsonify({
-                "success": True, 
-                "settings_dict": settings_payload
-            }), 200
-
-    except Exception as e:
-
-        short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-
-        log_action('Getting Settings', False, short_traceback, None, 'FAILED_GET_SETTINGS')
-
-        return jsonify({"success": False, "message": "There was an error whilst retrieving settings"}), 500
-
-@app.route("/save_settings", methods=["POST"]) 
-def save_settings():
-    data = request.get_json()
-    settings = data.get("settings_dict") or {} 
-    user = get_current_user()
-
-    if not user:
-        return jsonify({"success": False, "message": "Error: no user found"}), 401
-
-    try:
-        
-        with Session(engine) as db:
-            
-            # get existing records (if any)
-            existing_records = db.exec(
-                select(Settings)
-                .where(Settings.user_id == user.id)
-            ).all()
-
-            # this creates a dictionary of settings based on the exisiting records retrieved
-            settings_dictionary = {record.key: record for record in existing_records}
-
-            # tracker used to check if the db changed so that it can be determined whether to take any action or not
-            db_changed = False 
-
-            # this for loop goes through each setting in local storage and checks if it is the same in the db records
-            # it then sets the db_changed tracker value to true if there is change detected
-            # it also adds any new settings detected
-            for key, new_value in settings.items():
-                if key in settings_dictionary:
-                    record = settings_dictionary[key]
-                    if record.value != new_value:
-                        record.value = new_value
-                        db_changed = True
-                else:
-                    new_record = Settings(user_id=user.id, key=key, value=new_value)
-                    db.add(new_record)
-                    db_changed = True 
-            
-            # the db's changes are committed if any changes have been detected
-            if db_changed:
-                db.commit()
-            
-            return jsonify({"success": True, "message": "Successfully saved settings"}), 200
-
-    except Exception as e:
-        db.rollback()
-
-        short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-
-        log_action('Saving Settings', False, short_traceback, None, 'FAILED_SAVE_SETTINGS')
-
-        return jsonify({"success": False, "message": "There was an error whilst saving settings"}), 500
-#endregion
-
-service = NodeFinder(graph_path=Config.GRAPH_PATH)
-
 if os.getenv("LOAD_GRAPH_ON_IMPORT", "1").lower() not in ("0", "false", "no"):
     service.load_graph()
 
-
-
-
-def haversine(x1, y1, x2, y2):
-
-    # NOTE : coords need to be in lon/lat, make sure to convert coords
-
-    R = 6371000
-
-    phi1 = math.radians(y1)
-    phi2 = math.radians(y2)
-
-    dphi = math.radians(y2 - y1)
-    dlambda = math.radians(x2 - x1)
-
-    a = math.sin(dphi/2)**2 + math.cos(phi1)*math.cos(phi2)*math.sin(dlambda/2)**2
-    c = 2 * math.atan2(math.sqrt(a), math.sqrt(1-a))
-
-    return R * c
-
-def normalise_route(coordinates, avg_speed_kmh=4.5):
-    """
-    This function converts raw route geometry into metrics e.g ETA and elevation data (if present)
-    """
-
-    converted_coords = []
-
-    if not coordinates or len(coordinates) < 2:
-        return({"success": False, "message": "Route is too small"})
-    
-    start_coord = coordinates[0]
-
-    
-    # this checks the first coord to check if it is in lat/lon format, if not it converts coords to lat/lon format
-    if not (abs(start_coord[0]) <= 180 and abs(start_coord[1]) <= 90):
-        print("DEBUG STATEMENT\nFUNCTION : normalise_route()\nDetected projection that is NOT Lat/Lon for start point. Converting to Lat/Lon")
-        for coord in coordinates:
-            converted_coord = service.convert_web_mercator_to_wgs84(coord[0], coord[1])
-            converted_coords.append(converted_coord)
-
-    total_distance_m = 0.0
-    total_elevation_gain_m = 0.0
-
-    has_elevation = len(coordinates[0]) == 3
-
-    for i in range(len(coordinates) - 1):
-        x1, y1 = coordinates[i][0], coordinates[i][1]
-        x2, y2 = coordinates[i + 1][0], coordinates[i + 1][1]
-
-        # distance calculation
-        if not converted_coords or converted_coords == []:
-            total_distance_m += haversine(x1, y1, x2, y2)
-        else:
-            total_distance_m += haversine(converted_coords[i][0], converted_coords[i][1], converted_coords[i + 1][0], converted_coords[i + 1][1])
-
-        # elevation gain (only if available)
-        if has_elevation:
-            e1 = coordinates[i][2]
-            e2 = coordinates[i + 1][2]
-
-            if e2 > e1:
-                total_elevation_gain_m += (e2 - e1)
-
-    distance_km = total_distance_m / 1000.0
-    eta_hours = distance_km / avg_speed_kmh
-    eta_seconds = round(eta_hours * 60 * 60, 2)
-
-    return {
-        "distance_m": total_distance_m,
-        "distance_km": distance_km,
-        "elevation_gain_m": total_elevation_gain_m,
-        "eta_seconds": eta_seconds
-    }
-
-# helper function which returns true if the first coordinate has 3 values (i.e x, y and z)
-def check_elevation(coords):
-    return bool(coords) and len(coords[0]) == 3
-
-# returns the x and y coords of a coordinate that may be 2D or 3D
-def get_xy(coord):
-    if len(coord) >= 2:
-        return coord[0], coord[1]
-    return coord
-
-# if a coord is over 180 degrees then it is definitely not wgs84 (181 used as buffer for edge cases)
-# this helper thus returns true if a coord is web mercator and false if it isn't (and thus is wgs84)
-def check_web_mercator(coord):
-    if coord is None:
-        return False
-    return abs(coord[0]) > 181 or abs(coord[1]) > 181
-
-def parse_elevation(elevation_string: str):
-    if not elevation_string:
-        return None
-    try:
-        return float(elevation_string.replace('m', '').strip())
-    except Exception:
-        return None
-
-# helper function used in creation of gpx / geojson files which converts hrs and minutes to seconds 
-def parse_eta_to_seconds(eta_string: str):
-    if not eta_string:
-        return None
-    try:
-        hours = minutes = 0
-        if 'h' in eta_string:
-            hours = int(eta_string.split('h')[0].strip())
-        if 'm' in eta_string:
-            minutes_part = eta_string.split('h')[-1] if 'h' in eta_string else eta_string
-            minutes = int(minutes_part.replace('m', '').strip())
-        return hours * 3600 + minutes * 60
-    except Exception:
-        return None
-
-#region GEOSPATIAL FILE PROCESSING HELPER FUNCTIONS
-
-def generate_geojson(route):
-
-    raw_coordinates_str = route.coordinates
-
-    
-    # check to ensure that the coords are retrieved
-    if not raw_coordinates_str:
-        return None
-
-    # converts the raw coords into a list
-    raw_coords = json.loads(raw_coordinates_str)
-    
-    corrected_coords = []
-    
-    has_elevation = check_elevation(raw_coords)
-        
-    for coord in raw_coords:
-        if has_elevation:
-            x, y, elevation = coord
-            lon, lat = service.convert_web_mercator_to_wgs84(x, y)
-            corrected_coords.append([lon, lat, elevation])
-        else:
-            x, y = coord
-            lon, lat = service.convert_web_mercator_to_wgs84(x, y)
-            corrected_coords.append([x, y])
-
-        
-    # constructs the geojson dict
-    geojson_feature = {
-        "type": "Feature",
-        "geometry": {
-            "type": "LineString",
-            "coordinates": corrected_coords
-        },
-        "properties": {
-            "route_id": route.id,
-            "distance_km": route.distance_km,
-            "created_at": route.created_at.isoformat()
-        }
-    }
-    
-    # returns as JSON string
-    return json.dumps(geojson_feature)
-
-def generate_gpx(route):
-    gpx = gpxpy.gpx.GPX() # initialises container
-    track = gpxpy.gpx.GPXTrack(name=route.name) # creates a track (the entire route) and adds it to the container (next line)
-    gpx.tracks.append(track) 
-    segment = gpxpy.gpx.GPXTrackSegment() # creates a segment (one connected part of a route) and adds it to the container (next line)
-    track.segments.append(segment)
-
-    coords = json.loads(route.coordinates)
-
-    has_elevation = check_elevation(coords)
-
-    for coord in coords: # loops through each coord in one sub array in the coords array
-        if has_elevation:
-            x, y, elevation = coord
-            lon, lat = service.convert_web_mercator_to_wgs84(x, y) # converts to lat and lon (needed for gpx)
-            segment.points.append(gpxpy.gpx.GPXTrackPoint(
-                latitude=lat,
-                longitude=lon,
-                elevation=elevation
-            ))
-        else:
-            x, y = coord
-            lon, lat = service.convert_web_mercator_to_wgs84(x, y)
-            segment.points.append(gpxpy.gpx.GPXTrackPoint(
-                latitude=lat,
-                longitude=lon
-            ))
-
-    return gpx.to_xml()
-
-def parse_kml_coord_list(coord_list) -> list:
-    """
-    converts a list of KML coord tuples into the format of this app
-
-    KML format : (lon, lat) / (lon, lat, ele)
-
-    App format : [lat, lon] / [lat, lon, ele]
-    """
-
-    print(f"coord list: {coord_list}")
-
-    coords = []
-
-    for coord in coord_list:
-        lon = coord[0]
-        lat = coord[1]
-        ele = coord[2] if len(coord) > 2 else None
-        coords.append([lat, lon, ele]) if len(coord) > 2 else coords.append([lat, lon])
-    
-    return coords
-
-
-def process_kml_feature(feature) -> list:
-    """
-    this (recursively) processes KML features returns a list of coords extracted from any geometry found within KML features
-    """
-
-    extracted = [] 
-
-    print(f"feature : {feature}")
-
-    # if the feature has geometry
-    if hasattr(feature, "geometry") and feature.geometry:
-        geom = feature.geometry
-
-        print(geom)
-        
-
-        # this handles LineStrings
-        if geom.__class__.__name__ == "LineString":
-            extracted.extend(parse_kml_coord_list(geom.coords))
-            print(geom.coords)
-        
-        # this handles MultiLineStrings
-        if geom.__class__.__name__ == "MultiLineString":
-            for line in geom.geoms:
-                extracted.extend(parse_kml_coord_list(line.coords))
-    
-    # if the feature contains nested features
-    if hasattr(feature, "features"):
-        for sub_feature in feature.features():
-            extracted.extend(process_kml_feature(sub_feature))
-    
-    return extracted
-
-def extract_kml_coords(doc) -> list:
-    """
-    this parses a KML file as text and extracts all coordinates from LineString and MultiLineString geometries
-
-    returns a coords list of [lat, lon, ele] points
-    """    
-
-
-    print(doc)
-    print(doc.features)
-    print(list(doc.features))
-
-    coords = []
-
-    # kml features can include documents, folders, or placemarks
-    for feature in doc.features:
-        print(feature)
-        coords.extend(process_kml_feature(feature))
-    
-    return coords
-
-
-#endregion
-
-# route which forms a path from the set of points the back-end receives as coordinates and returns it to the front-end
-@app.route("/calculate_path", methods=["POST"])
-@limiter.limit("4 per second")
-def calculate_path():
-    try:
-        data = request.get_json()
-        
-        # this extracts and parses start + end point coordinates
-        start_coords = data.get("start_point", "")
-        end_coords = data.get("end_point", "")
-
-        graph = service.load_graph()
-        
-        # Validation of coord format 
-        if not start_coords or not end_coords:
-            raise ValueError("Start and end coordinates are required")
-
-        if len(start_coords) < 2 or len(end_coords) < 2:
-            raise ValueError("Coordinates must be in format 'x, y'")
-
-        # this extracts the raw numerical coordinates
-        start_coords_x, start_coords_y = get_xy(start_coords)
-        end_coords_x, end_coords_y = get_xy(end_coords)
-
-        all_coords = [start_coords_x, start_coords_y, end_coords_x, end_coords_y]
-        if not all(isinstance(num, (int, float)) for num in all_coords):
-            raise ValueError("Coordinates must be valid numbers")
-        
-        s_x, s_y = start_coords_x, start_coords_y 
-        e_x, e_y = end_coords_x, end_coords_y 
-        
-        
-        # COORD PROJECTION TYPE DETECTION
-        # Lat / Lon coordinates will be small numbers, whereas web mercator uses large values in metres
-    
-        # this checks start and end coords to see if they are wgs84 projection and converts to web_mercator if so
-        if abs(s_x) <= 180 and abs(s_y) <= 90:
-            print("Detected Lat/Lon for start point. Converting to Web Mercator...")
-            s_x, s_y = service.convert_wgs84_to_web_mercator(s_x, s_y)
-        if abs(e_x) <= 180 and abs(e_y) <= 90:
-            print("Detected Lat/Lon for end point. Converting to Web Mercator...")
-            e_x, e_y = service.convert_wgs84_to_web_mercator(e_x, e_y)
-        
-    except (KeyError, ValueError) as e:
-        with Session(engine) as db:
-            user = get_current_user()
-            if user:
-                available_routes = db.exec(
-                    select(Route)
-                    .where(Route.user_id == user.id)
-                ).all()
-            else:
-                available_routes = []
-
-            short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-
-            log_action('Calculating Path', False, short_traceback, None, 'INVALID_COORDS_AUTO_PATH_CREATION')
-
-        return jsonify({
-            "success": False,
-            "map_centre": DEFAULT_CENTRE,
-            "available_routes": available_routes, 
-            "message": f"Invalid coordinates: {str(e)}"
-        })
-
-    # Build the route using the Web Mercator coordinates
-    path, start_node, end_node, time_taken = service.build_route(s_x, s_y, e_x, e_y)
-
-    if not path:  
-        with Session(engine) as db:
-            user = get_current_user()
-            if user:
-                available_routes = db.exec(
-                    select(Route)
-                    .where(Route.user_id == user.id)
-                ).all()
-            else:
-                available_routes = []
-                
-
-            log_action('Calculating Path', False, 'Path not created', None, 'NO_PATH_FOUND')
-            
-            return jsonify({
-                "success": False,
-                "map_centre": DEFAULT_CENTRE,
-                "available_routes": available_routes,
-                "message": "No path could be created"
-            })
-
-    web_mercator_coordinates = [] 
-    
-    # Since the graph/path is already in Web Mercator, we no longer need to convert it!
-    for node in path:  
-        x, y = node  
-        elev = graph.nodes.get(node, {}).get('elev')
-        if elev is not None:
-            web_mercator_coordinates.append([x, y, elev])
-        else:
-            web_mercator_coordinates.append([x, y])
-
-    elevation_gain = 0
-    elevation_change = 0
-    
-    # this calculates the elevation GAIN
-    for node1, node2 in zip(path, path[1:]):
-        elev1 = graph.nodes.get(node1, {}).get('elev')
-        elev2 = graph.nodes.get(node2, {}).get('elev')
-
-        if elev1 is not None and elev2 is not None:
-            elevation_gain += max(0, elev2 - elev1)
-
-    # this calculates the elevation CHANGE
-    start_elevation = int(graph.nodes[start_node]['elev'])
-    end_elevation = int(graph.nodes[end_node]['elev'])
-
-    elevation_change = end_elevation - start_elevation
-    
-    start_coords = web_mercator_coordinates[0]
-    end_coords = web_mercator_coordinates[-1]
-    
-    # Calculates distance and eta statistics
-    total_distance = service.calculate_route_distance(path)
-    
-    # Since your network graph distance calculation is now in meters (or handled by your service),
-    # we convert it directly to kilometers here
-    total_distance_km = total_distance / 1000
-    
-    # Calculates ETA 
-    # NOTE : this is in SECONDS
-    eta_seconds = service.calculate_eta(path, graph)
-
-    path_geojson = {
-        "type": "Feature",
-        "geometry": {"type": "LineString", "coordinates": web_mercator_coordinates},
-        "properties": {"color": "#2563eb"}
-    } 
-    
-    # Calculates optimal map centre and zoom
-    map_centre, map_zoom = service.calculate_map_center_and_zoom(web_mercator_coordinates)
-
-    route_stats = {
-        "start_elevation": start_elevation,
-        "end_elevation": end_elevation,
-        "elevation_change": elevation_change,
-        "elevation_gain": elevation_gain, 
-        "total_distance": round(total_distance_km, 2),
-        "eta_seconds": eta_seconds
-    }
-
-    with Session(engine) as db:
-
-        log_action('Calculating Path', True, f"distance: {round(total_distance_km, 2)}km", time_taken, 'PATH_CREATED')
-
-    user = get_current_user()
-    if user:
-        with Session(engine) as db:
-            available_routes = db.exec(
-                select(Route)
-                .where(Route.user_id == user.id)
-            ).all()
-    else:
-        available_routes = []
-            
-    return jsonify({
-        "success": True,
-        "pathGeoJSON": path_geojson,
-        "map_centre": map_centre,
-        "map_zoom": map_zoom,
-        "route_stats": route_stats,
-        "coordinates": web_mercator_coordinates,
-        "startCoord": start_coords,
-        "endCoord": end_coords
-    })
-
-# route which saves a user-chosen point on the map
-@app.route("/save_point", methods=["POST"])
-@limiter.limit("10 per minute")
-def save_point():
-    # data received from front-end is broken down into web mercator coords that can be converted into bng
-    data = request.get_json()
-    point_name = data.get('point_name', '').strip()
-    web_mercator_x = data.get("web_mercator_x")
-    web_mercator_y = data.get("web_mercator_y")
-
-    if not point_name or web_mercator_x is None or web_mercator_y is None:
-        return jsonify({"success": False, "message": "Name and Coordinates are required"})
-    
-    try:
-              
-        # coords are used to save the chosen point
-        coords = json.dumps([float(web_mercator_x), float(web_mercator_y)])
-
-        user = get_current_user()
-        if user:
-            with Session(engine) as db:
-                new_point = Point(name=point_name, coordinates=coords, user_id=user.id) 
-                db.add(new_point)
-                db.commit()
-
-                log_action('Saving Point', True, None, None, 'SAVING_POINT')
-
-            return jsonify({"success": True, "message": 'Successfully saved the point'})
-    
-    # if float() fails
-    except ValueError as e:
-
-        short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()     
-        log_action('Saving Point', False, short_traceback, None, 'SAVING_POINT')
-
-        return jsonify({"success": False, "message": "Invalid coordinate format. Coordinates must be numbers."}), 400
-        
-    except Exception as e:
-
-        with Session(engine) as db: 
-
-            short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-        
-            log_action('Saving Point', False, short_traceback, None, 'SAVING_POINT')
-
-        return jsonify({"success": False, "message": "Failed to save point."}), 500
-
-
-# flask-route which is used to save a route that has been passed into the backend into the PostgreSQL database 
-@app.route("/save_route", methods=["POST"])
-@limiter.limit("110 per minute")
-def save_route():
-    with Session(engine) as db:
-        try:
-            data = request.get_json()
-            if not data:
-                return jsonify({"success": False, "message": "Invalid request"}), 400
-
-            user = get_current_user()
-
-            route_name = data.get("route_name")
-            coordinates = data.get("coordinates")
-            type = data.get("type")
-
-            if not route_name or not route_name.strip():
-                return jsonify({"success": False, "message": "Route name required"}), 400
-            
-
-            if not coordinates or len(coordinates) < 2:
-                return jsonify({"success": False, "message": "Invalid route data"}), 400
-            
-            metrics = normalise_route(coordinates)
-
-            route = Route(
-                name=route_name,
-                coordinates=json.dumps(coordinates),
-                user_id=user.id,
-
-                distance_km=metrics["distance_km"],
-                eta_seconds=metrics["eta_seconds"],
-                elevation_change=metrics["elevation_gain_m"]
-            )
-
-            route_info = {
-                "route_name": route_name,
-                "distance_km": metrics["distance_km"],
-                "eta_seconds": metrics["eta_seconds"],
-                "elevation_gain_metres": metrics["elevation_gain_m"],
-            }
-
-            
-            db.add(route)
-            db.commit()
-
-            log_action('Saving Route', True, None, None, 'SAVE_ROUTE')
-
-            return jsonify({
-                "success": True,
-                "route_info": route_info
-            })
-
-
-        except IntegrityError as e:
-            db.rollback()
-
-            short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-
-            log_action('Saving Route', False, short_traceback, None, 'SAVE_ROUTE')
-
-            return jsonify({"success" : False, "message" : "There was an error saving your route. "})
-
-
-        except Exception as e:
-            db.rollback()
-
-            short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-            
-            log_action('Saving Route', True, short_traceback, None, 'SAVE_ROUTE')
-
-            return jsonify({"success": False, "message": "There was an error saving your route. "})
-    
-
-
-# flask route which is used to retrieves a saved route that has been passed into the back-end from the PostgreSQL database
-@app.route("/load_route", methods=["POST"])
-@limiter.limit("110 per minute")
-def load_route():
-    data = request.get_json()
-    route_name = data.get('route_name', '').strip()
-    
-    if not route_name:
-        return jsonify({"success": False, "message": "Route name is required"})
-    
-    with Session(engine) as db:
-        route = db.exec(
-            select(Route)
-            .where(Route.name == route_name)
-        ).first()
-
-        if not route:
-            return jsonify({"success": False, "message": "Route not found"})
-        
-        json_coords = json.loads(route.coordinates)
-
-        
-        # success, coordinates, file_type = db_manager.load_route(route_name)
-        coordinates = json_coords
-
-        has_elevation = check_elevation(coordinates)
-        
-        try:
-            # converts wgs84 coordinates to web mercator for display
-            web_mercator_coordinates = []
-
-            if check_web_mercator(coordinates[0]):
-                # already in web mercator so this block just normalises the lengths
-                for coord in coordinates:
-                    x, y = coord[0], coord[1]
-                    z = coord[2] if len(coord) >= 3 else 0   # default elevation = 0
-                    web_mercator_coordinates.append([x, y, z])
-            else:
-                # this block converts wgs84 to web mercator
-                for coord in coordinates:
-                    if has_elevation and len(coord) >= 3:
-                        lon, lat, elevation = coord[:3]
-                        web_x, web_y = service.convert_wgs84_to_web_mercator(lon, lat)
-                        web_mercator_coordinates.append([web_x, web_y, elevation])
-                    else:
-                        lon, lat = coord[:2]
-                        web_x, web_y = service.convert_wgs84_to_web_mercator(lon, lat)
-                        web_mercator_coordinates.append([web_x, web_y, 0])
-
-            if not web_mercator_coordinates:
-
-                log_action('Loading Route', False, "No coordinates found in the route", None, 'LOAD_ROUTE')
-
-                return jsonify({"success": False, "message": "No valid coordinates found in route file"})
-            
-            # converts coordinates to geojson format for display
-            path_geojson = {
-                "type": "FeatureCollection",
-                "features": [{
-                    "type": "Feature",
-                    "geometry": {"type": "LineString", "coordinates": web_mercator_coordinates},
-                    "properties": {"color": "#2563eb"}
-                }]
-            }
-            
-            
-            # calculates midpoint for map centring
-            midpoint = web_mercator_coordinates[len(web_mercator_coordinates)//2] if web_mercator_coordinates else DEFAULT_CENTRE
-
-            # data collected directly from database values
-            eta_seconds = route.eta_seconds
-            distance = route.distance_km
-            elevation_change = route.elevation_change
-            
-            # route statistics to pass to frontend
-            route_stats = {
-                "total_distance": distance if distance is not None else 0,
-                "eta_seconds": eta_seconds,
-                "elevation_change": elevation_change
-            }
-
-            log_action('Loading Route', True, None, None, 'LOAD_ROUTE')
-            
-            return jsonify({
-                "success": True, 
-                "message": f"Route '{route_name}' loaded successfully",
-                "pathGeoJSON": path_geojson,
-                "map_centre": midpoint,
-                "coordinates": web_mercator_coordinates,
-                "route_stats": route_stats
-            })
-        
-        except Exception as e:
-
-            short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-            
-            log_action('Loading Route', False, short_traceback, None, 'LOAD_ROUTE')
-
-            return jsonify({"success": False, "message": "The route could not be loaded"})
-
-
-# flask route which returns a raw binary file of the route in either GPX or GeoJSON format
-@app.route("/download_route", methods=["POST"])
-def download_route():
-    data = request.get_json(silent=True) or {} # silent to prevent errors and return None
-    
-    route_name = data.get('route_name', "").strip()
-    file_type = data.get('route_type').lower()
-
-    # ensures that routes have a name (defensive programming since it is almost impossible a route doesn't have a name due to design of app)
-    if not route_name:
-        return jsonify({"success": False, "message": "Route name is required"}), 400
-
-    # ensures that routes have a file type (also impossible since download occurs through either a gpx or a geojson button)
-    if file_type not in ["gpx", "geojson"]:
-        return jsonify({"success": False, "message": "Invalid file type. Use 'gpx' or 'geojson'"}), 400
-
-    user = get_current_user()
-    # ensures that a user is logged in
-    if not user:
-        return jsonify({"success": False, "message": "You must be logged in"}), 401
-
-    with Session(engine) as db:
-
-        route = db.exec(
-            select(Route)
-            .where(Route.name == route_name, Route.user_id == user.id)
-        ).first()
-        
-        # ensures that a route is found before attempting to convert
-        if not route:
-
-            log_action('Downloading Route', False, 'Route not found', None, 'DOWNLOAD_ROUTE')
-
-            return jsonify({"success": False, "message": "Route not found or you don't own it"}), 404
-
-        print(route.coordinates)
-
-        try:
-            if file_type == "gpx":
-                content = generate_gpx(route)
-                mimetype = 'application/gpx+xml'
-                extension = 'gpx'
-            else:
-                content = generate_geojson(route)
-                mimetype = 'application/geo+json'
-                extension = 'geojson'
-
-            safe_name = "".join(
-                char if char.isalnum() or char in " -_()" else "_" 
-                for char in route.name
-            )
-
-            log_action('Downloading Route', True, extension, None, 'DOWNLOAD_ROUTE')
-
-            return send_file(
-                io.BytesIO(content.encode('utf-8')),
-                mimetype=mimetype,
-                as_attachment=True,
-                download_name=f"{safe_name}.{extension}"
-            )
-
-        except Exception as e:
-
-            short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-
-            log_action('Downloading Route', False, short_traceback, None, 'DOWNLOAD_ROUTE')
-
-            return jsonify({"success": False, "message": "Failed to download route"}), 500
-
-@app.route("/import_route_file", methods=["POST"])
-def import_route():
-    """
-    This function handles route file uploads for GPX, FIT, KML, and GeoJSON formats.
-
-    it received as parameters:
-        - route file
-
-    it returns:
-        JSON containing:
-            success (bool)
-            coords (list of [lat, lon, ele?])
-    """
-
-    # this retrieves the uploaded file from the request
-    uploaded_file = request.files.get("route_file")
-
-    if not uploaded_file:
-
-        log_action('Importing Route', False, 'Uploaded file could not be retrieved', None, 'IMPORT_ROUTE')
-
-        return jsonify({"success": False, "message": "Cannot receive uploaded file"}), 400
-    
-    try: 
-    
-        # this extracts the filename and file extension
-        filename = uploaded_file.filename
-        ext = filename.rsplit('.', 1)[-1].lower()
-
-        # this validates the supported formats 
-        if ext not in ["gpx", "fit", "kml", "geojson"]:
-            return jsonify({"success": False, "message": "Please upload a file of the supported types"}), 400
-        
-        # this reads raw bytes (works for both binary + text formats) for FIT file type
-        raw = uploaded_file.read()
-
-        # this decodes text formats (GPX, KML, GeoJSON) 
-        text = raw.decode("utf-8", errors="ignore")
-
-        # this handles GPX file types
-        if ext == "gpx":
-            # Parse GPX XML
-            gpx = gpxpy.parse(text)
-
-            converted_points = []
-
-            # Extract all track points into [lat, lon, ele]
-            points = [
-                [p.latitude, p.longitude, p.elevation]
-                for t in gpx.tracks
-                for s in t.segments
-                for p in s.points
-            ]
-
-            for point in points:
-                
-                lat, lon, ele = point[0], point[1], point[2]
-
-                web_mercator_coords = service.convert_wgs84_to_web_mercator(lon, lat)
-
-                converted_points.append([web_mercator_coords[0], web_mercator_coords[1], ele])
-
-            return jsonify({
-                "success": True,
-                "coords": converted_points,
-            })
-
-        # this handles FIT file types
-        elif ext == "fit":
-            coords = []
-            fitfile = FitFile(raw)
-
-            for record in fitfile.get_messages("record"):
-                lat = record.get_value("position_lat")
-                lon = record.get_value("position_long")
-
-                # Skip missing coordinate records
-                if lat is None or lon is None:
-                    continue
-
-                # FIT stores coordinates in semicircles → convert to degrees
-                lat_deg = lat * (180 / 2**31)
-                lon_deg = lon * (180 / 2**31)
-
-                coords.append([lat_deg, lon_deg])
-
-            return jsonify({"success": True, "coords": coords})
-
-        # this handles KML file types
-        elif ext == "kml":
-            uploaded_file.stream.seek(0)
-            doc = KML.parse(uploaded_file.stream, strict=False)
-            coords = extract_kml_coords(doc)
-            return jsonify({"success": True, "coords": coords})
-
-        # this handles geojson file types
-        elif ext == "geojson":
-            geo = json.loads(text)
-            coords = []
-            geometries = []
-
-            """
-            
-            structure of geojson files is usually like the below: 
-
-            {{
-                "type": "FeatureCollection",
-                    "features": [
-                        {
-                        "type": "Feature",
-                        "geometry": {
-                            "type": "LineString",
-                            "coordinates": [[-1.8, 53.3, 250], [-1.8, 53.4, 260]]
-                        },
-                        "properties": {}
-                        }
-                    ]
-                } 
-            """
-
-            # this extracts the object allowing us to use conditional logic to work towards the coords of the file
-            root_type = geo.get("type")
-
-            if root_type == "FeatureCollection":
-                for feature in geo.get("features", []):
-                    if "geometry" in feature:
-                        geometries.append(feature["geometry"])
-            elif root_type == "Feature":
-                if "geometry" in geo:
-                    geometries.append(geo["geometry"])
-            else:
-                # this is if the root is geometry itself
-
-                # validate that the geom type is actually a geometry 
-                if geo.get('type') in ["Point", "LineString", "MultiLineString", "Polygon"]:
-                    geometries.append(geo)
-                else:
-                    log_action('Importing Route', False, 'Invalid GeoJSON Structure', None, "IMPORT_ROUTE")
-                    return jsonify({"success": False, "message": "There was an error on our end, please try again later."})
-
-            # this is for processing LineString geometries
-            for geom in geometries:
-                if geom and geom.get("type") == "LineString":
-                    for coord in geom.get("coordinates", []):
-                        # this unpacks the values from each coord
-
-                        if len(coord) <= 1:
-                            return jsonify({"success": False, "message": "Error whilst parsing GeoJSON: given coordinates have one value only"})
-
-                        lon = coord[0]
-                        lat = coord[1]
-                        ele = coord[2] if len(coord) > 2 else 0
-
-                        web_mercator_x, web_mercator_y = service.convert_wgs84_to_web_mercator(lon, lat)
-                        coords.append([web_mercator_x, web_mercator_y, ele])
-            
-            log_action('Importing Route', True, ext, None, 'IMPORT_ROUTE')
-
-            return jsonify({"success": True, "coords": coords})
-    except Exception as e:
-            
-            short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-
-            log_action('Importing Route', False, short_traceback, None, 'IMPORT_ROUTE')
-            return jsonify({"success": False, "message": "There was an error on our end, please try again later."})
-
-
-
-# flask-route which retrieves saved points to be used in the front
-@app.route("/get_saved_points", methods=["GET"])
-def get_saved_points():
-
-    user = get_current_user()
-    if user:
-        with Session(engine) as db:
-            points = db.exec(
-                select(Point)
-                .where(Point.user_id == user.id)
-            ).all()
-    
-    if not points:
-        return jsonify({"success": False, "message": "No points found" ,"points": []})
-    
-    web_mercator_points = []
-    for point in points:
-        try:
-            web_mercator_x, web_mercator_y = json.loads(point.coordinates)
-
-            web_mercator_points.append({
-                "name": point.name,
-                "coordinates": [web_mercator_x, web_mercator_y]
-            })
-        except Exception as e:
-            
-            short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-
-            log_action('Getting Saved Points', False, short_traceback, None, 'GET_SAVED_POINT')
-    
-    
-    return jsonify({"success": True, "points": web_mercator_points})
-
-# route which deletes a saved point that is passed into the back-end from the front-end
-@app.route("/delete_point", methods=['POST'])
-@limiter.limit("110 per minute")
-def delete_point():
-    data = request.get_json()
-    point_name = data.get('point_name', '').strip()
-
-    if not point_name:
-        return jsonify ({"success": False, "message": "Point name is missing"})
-    
-    try:
-        with Session(engine) as db:
-
-            point_to_delete = db.exec(
-                select(Point)
-                .where(Point.name == point_name)
-            ).first()
-
-            db.delete(point_to_delete)
-            db.commit()
-
-            log_action('Deleting Point', True, None, None, 'DELETE_POINT')
-
-            return jsonify({"success": True, "message": f"Successfully deleted the {point_name} point"})
-
-    except Exception as e:
-
-        short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-        
-        log_action('Deleting Point', False, short_traceback, None, 'DELETE_POINT')
-
-        return jsonify({"success": False, "message": f"Could not successfully save the {point_name} point"})
-    
-    
-#flask-route which deletes a route that is passed from the front-end to the back-end
-@app.route("/delete_route", methods=["POST"])
-@limiter.limit("110 per minute")
-def delete_route():
-    data = request.get_json()
-    route_name = data.get("route_name")
-    user = get_current_user()
-
-    if not route_name:
-        return jsonify({"success": False, "message": "Route name is missing"}), 400
-
-    try:
-        with Session(engine) as db:
-            route = db.exec(
-                select(Route)
-                .where(Route.name == route_name, Route.user_id == user.id)
-            ).first()
-            db.delete(route)
-            db.commit()
-
-            return jsonify({"success": True, "message": f"Successfully saved the {route.name}"})
-    except Exception as e:
-
-        short_traceback = "".join(traceback.format_exception_only(type(e), e)).strip()
-        
-        log_action('Deleting Route', False, short_traceback, None, 'DELETE_ROUTE')
-
-        return jsonify({"success": False, "message": "Could not delete the route."})
-
-
-# GENERAL MAP ROUTES
+#region FLASK ROUTES: VIEW ROUTERS
 
 @app.route('/')
 @limiter.exempt
@@ -1522,7 +163,10 @@ def register_page():
 
     return render_template("register.html")
 
-# After logging in
+#endregion
+
+#region FLASK ROUTE: MAIN /MAP ROUTE
+
 @app.route("/map")
 @limiter.exempt
 def map_view():
@@ -1574,6 +218,9 @@ def map_view():
 
             return jsonify({"success": False, "message": f"Error whilst getting map: {e}"})
 
+#endregion
+
+#region FLASK ROUTE: SEARCHING
 
 @app.route("/search_area", methods=["POST"])
 @limiter.limit("10 per minute")
@@ -1618,7 +265,7 @@ def search_area():
     except Exception:
         log_action('Searching for Area', False, traceback.format_exc(), None, 'SEARCH_AREA')
 
-
+#endregion
 
 
 if __name__ == "__main__":
