+++ src/config.py
@@ -16,4 +16,4 @@ class Config:
     LOCATIONIQ_API_KEY = os.getenv('LOCATIONIQ_API_KEY')
     DATABASE_URI = os.getenv('DATABASE_URI')
     LOCAL_DATABASE_URI = os.getenv('LOCAL_DATABASE_URI')
-    GRAPH_PATH = os.getenv('GRAPH_PATH', 'Pathfinding/better_path_graph.pkl')
+    GRAPH_PATH = os.getenv('GRAPH_PATH')
