+++ templates/map.html
@@ -474,17 +474,17 @@
                 {% endfor %}
             {% endif %}
             },
-            apiLoginUrl: "{{ url_for('login') }}",
-            apiLogoutUrl: "{{ url_for('logout') }}",
-            apiRegisterUrl: "{{ url_for('registering') }}",
+            apiLoginUrl: "{{ url_for('auth_api.login') }}",
+            apiLogoutUrl: "{{ url_for('auth_api.logout') }}",
+            apiRegisterUrl: "{{ url_for('auth_api.registering') }}",
             apiCalculatePathUrl: "{{ url_for('route_api.calculate_path') }}",
             apiSearchAreaUrl: "{{ url_for('search_area') }}",
-            apiSavePointUrl: "{{ url_for('save_point') }}",
-            apiGetSavedPointsUrl: "{{ url_for('get_saved_points') }}",
-            apiDeletePointUrl: "{{ url_for('delete_point') }}",
+            apiSavePointUrl: "{{ url_for('point_api.save_point') }}",
+            apiGetSavedPointsUrl: "{{ url_for('point_api.get_saved_points') }}",
+            apiDeletePointUrl: "{{ url_for('point_api.delete_point') }}",
             apiDeleteRouteUrl: "{{ url_for('route_api.delete_route') }}",
             apiSaveRouteUrl: "{{ url_for('route_api.save_route') }}",
-            apiDeleteAccountUrl: "{{ url_for('delete_account') }}",
+            apiDeleteAccountUrl: "{{ url_for('auth_api.delete_account') }}",
             apiLoadRouteUrl: "{{ url_for('route_api.load_route') }}",
             mapPageUrl: "{{ url_for('map_view') }}",
             apiDownloadRouteUrl: "{{ url_for('route_api.download_route') }}",
