+++ templates/register.html
@@ -60,6 +60,6 @@
 </body>
 <script type="module" src="{{ url_for('static', filename='js/auth.js') }}"></script>
 <script>
-    const apiRegisterUrl = "{{ url_for('registering') }}";
+    const apiRegisterUrl = "{{ url_for('auth_api.registering') }}";
 </script>
 </html>
